Rename badly named statser package
