+++ pkg/backends/newrelic/newrelic.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/atlassian/gostatsd"
-	stats "github.com/atlassian/gostatsd/pkg/statser"
+	"github.com/atlassian/gostatsd/pkg/stats"
 
 	"github.com/cenkalti/backoff"
 	log "github.com/sirupsen/logrus"
