+++ pkg/stats/flush_notifier.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"sync"
