+++ pkg/stats/flush_notifier_test.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"testing"
