+++ pkg/stats/heartbeat.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"context"
