+++ pkg/stats/statser_internal.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"context"
