+++ pkg/stats/statser_null.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"time"
