+++ pkg/stats/statser_tagged.go
@@ -1,4 +1,4 @@
-package statser
+package stats
 
 import (
 	"time"
