+++ pkg/statsd/handler_cloud.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/atlassian/gostatsd"
-	stats "github.com/atlassian/gostatsd/pkg/statser"
+	"github.com/atlassian/gostatsd/pkg/stats"
 
 	"github.com/ash2k/stager/wait"
 	"github.com/sirupsen/logrus"
